avanzo con la segregacion
