+++ funcion/entrarDiariamente.py
@@ -3,9 +3,9 @@
 
 import time
 import traceback
-from bot.botComm import *
-from sisop.varios import *
-from sisop.signals import *
+from helpers.botComm import *
+from helpers.signals import *
+from helpers.gui import *
 
 def entrarDiariamente(s):
 	if botValido(s) is False:
