+++ funcion/getStatus.py
@@ -4,11 +4,11 @@
 import re
 from decimal import *
 from config import *
-from sisop.varios import *
 from helpers.varios import *
 from helpers.getJson import *
 from helpers.naval import *
 from helpers.recursos import *
+from helpers.gui import *
 
 getcontext().prec = 30
