+++ helpers/process.py
@@ -1,0 +1,21 @@
+#! /usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+import os
+import subprocess
+
+def forkear(s):
+	newpid = os.fork()
+	if newpid != 0:
+		# padre
+		s.login()
+		newpid = str(newpid)
+		run('kill -SIGSTOP ' + newpid)
+		run('bg ' + newpid)
+		run('disown ' + newpid)
+	else:
+		# hijo
+		s.padre = False
+
+def run(command):
+	return subprocess.Popen(command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE).stdout
